+++ src/TVCS/Utils/ToonPoint.java
@@ -6,8 +6,8 @@ import java.io.Serializable;
  * Created by ina on 2017-06-05.
  */
 public class ToonPoint implements Serializable {
-    public int x, y;
-    public ToonPoint(int x, int y) {
+    public double x, y;
+    public ToonPoint(double x, double y) {
         this.x = x;
         this.y = y;
     }
